+++ sdk/src/core/Injectable.ts
@@ -291,8 +291,8 @@ import { UpdateConfigVersionCommandHandler } from '../app/usecase/command/manage
 import { UpdateResolverCommandHandler } from '../app/usecase/command/management/updateResolver/updateResolverCommandHandler.js';
 import { GetConfigInfoQueryHandler } from '../app/usecase/query/management/GetConfigInfoQueryHandler';
 import { UpdateMaturityDateCommandHandler } from '../app/usecase/command/bond/updateMaturityDate/UpdateMaturityDateCommandHandler.js';
-import { SetScheduledBalanceAdjustmentCommandHandler } from '../app/usecase/command/equity/setScheduledBalanceAdjustment/SetScheduledBalanceAdjustmentCommandHandler.js';
-import { GetScheduledBalanceAdjustmentQueryHandler } from '../app/usecase/query/equity/getScheduledBalanceAdjustment/GetScheduledBalanceAdjustmentQueryHandler.js';
+import { SetScheduledBalanceAdjustmentCommandHandler } from '../app/usecase/command/equity/balanceAdjustments/setScheduledBalanceAdjustment/SetScheduledBalanceAdjustmentCommandHandler.js';
+import { GetScheduledBalanceAdjustmentQueryHandler } from '../app/usecase/query/equity/balanceAdjustments/getScheduledBalanceAdjustment/GetScheduledBalanceAdjustmentQueryHandler.js';
 
 export const TOKENS = {
   COMMAND_HANDLER: Symbol('CommandHandler'),
